requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
sqlparse==0.5.0
